# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ParkEase Nairobi - Peer-to-peer parking marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: spotId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled, no_show]
        - in: query
          name: role
          schema:
            type: string
            enum: [commuter, owner]
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingList"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created (payment pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking (access instructions revealed if confirmed)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (confirm payment, cancel, mark no-show)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─── REVIEWS ──────────────────────────────────────────────────────────
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - in: query
          name: spotId
          schema:
            type: integer
        - in: query
          name: revieweeId
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── WAITLIST ─────────────────────────────────────────────────────────
...
  /analytics/owner-dashboard:
    get:
      operationId: getOwnerDashboard
      tags: [analytics]
      summary: Owner dashboard summary (earnings, occupancy, bookings)
      parameters:
        - in: query
          name: ownerId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerDashboard"
...
  /analytics/surge:
    get:
      operationId: getSurgePricing
      tags: [analytics]
      summary: Surge pricing multiplier for a zone and date
      parameters:
        - in: query
          name: zone
          required: true
          schema:
            type: string
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Surge pricing info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurgePricing"
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        spotId:
          type: integer
        spotTitle:
          type: string
        spotAddress:
          type: string
        spotLat:
          type: number
        spotLng:
          type: number
        comuterId:
          type: integer
        commuterName:
          type: string
        commuterPhone:
          type: string
        ownerId:
          type: integer
        date:
          type: string
          format: date
        startHour:
          type: integer
        endHour:
          type: integer
        totalHours:
          type: number
        pricePerHour:
          type: number
        surgeMultiplier:
          type: number
        totalAmount:
          type: number
        platformFee:
          type: number
        ownerEarning:
          type: number
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        mpesaCode:
          type: string
          nullable: true
        accessInstructions:
          type: string
          nullable: true
          description: Only returned when status is confirmed or completed
        cancellationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, spotId, spotTitle, spotAddress, spotLat, spotLng, comuterId, commuterName, commuterPhone, ownerId, date, startHour, endHour, totalHours, pricePerHour, surgeMultiplier, totalAmount, platformFee, ownerEarning, status, createdAt]
    BookingList:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
      required: [bookings, total]
    CreateBookingBody:
      type: object
      properties:
        spotId:
          type: integer
        commuterId:
          type: integer
        date:
          type: string
          format: date
        startHour:
          type: integer
        endHour:
          type: integer
        mpesaPhone:
          type: string
      required: [spotId, commuterId, date, startHour, endHour]
    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cancelled, completed, no_show]
        mpesaCode:
          type: string
        cancellationReason:
          type: string
      required: [status]

    # ─── REVIEW ────────────────────────────────────────────────────────
    Review:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        reviewerAvatarUrl:
          type: string
          nullable: true
        revieweeId:
          type: integer
          nullable: true
        spotId:
          type: integer
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, bookingId, reviewerId, reviewerName, rating, createdAt]
    ReviewList:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        averageRating:
          type: number
          nullable: true
      required: [reviews, total]
    CreateReviewBody:
      type: object
      properties:
        bookingId:
          type: integer
        reviewerId:
          type: integer
        revieweeId:
          type: integer
        spotId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [bookingId, reviewerId, rating]

    # ─── WAITLIST ──────────────────────────────────────────────────────
...
    OwnerDashboard:
      type: object
      properties:
        totalEarnings:
          type: number
        pendingPayouts:
          type: number
        totalBookings:
          type: integer
        completedBookings:
          type: integer
        cancelledBookings:
          type: integer
        occupancyRate:
          type: number
          description: Percentage 0-100
        activeSpots:
          type: integer
        averageSpotRating:
          type: number
          nullable: true
        earningsByMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyEarning"
        topSpots:
          type: array
          items:
            $ref: "#/components/schemas/SpotSummary"
      required: [totalEarnings, pendingPayouts, totalBookings, completedBookings, cancelledBookings, occupancyRate, activeSpots, earningsByMonth, topSpots]
    MonthlyEarning:
      type: object
      properties:
        month:
          type: string
        earnings:
          type: number
        bookings:
          type: integer
      required: [month, earnings, bookings]
    SpotSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        zone:
          type: string
        earnings:
          type: number
        bookings:
          type: integer
        rating:
          type: number
          nullable: true
      required: [id, title, zone, earnings, bookings]
...
    SurgePricing:
      type: object
      properties:
        zone:
          type: string
        date:
          type: string
          format: date
        multiplier:
          type: number
        reason:
          type: string
          nullable: true
        isEventDay:
          type: boolean
      required: [zone, date, multiplier, isEventDay]